refactor(split-rendering): quarantine all orchestration (Phase 3 done)
Complete Phase 3 by moving the remaining orchestration-layer code into
orchestration/:

- orchestration/render_buffer.rs: compute_buffer_layout,
  draw_buffer_in_split, render_buffer_in_split, resolve_cursor_fallback,
  BufferLayoutOutput
- orchestration/render_composite.rs: render_composite_buffer,
  render_view_line_content

The orchestration/ subdirectory is now the *only* place that imports
SelectionContext / DecorationContext. The 11 top-level self-contained
modules (spans, style, char_style, post_pass, base_tokens, transforms,
view_data, folding, scrollbar, layout, gutter) remain free of any shared
mega-struct dependency.

mod.rs shrinks from 8,635 lines (pre-refactor) to 2,734 lines.
All 37 split_rendering tests pass.
