+++ crates/fresh-editor/src/view/ui/split_rendering/mod.rs
@@ -15,6 +15,11 @@ mod view_data;
 
 use orchestration::contexts::{DecorationContext, SelectionContext};
 use orchestration::overlays::{decoration_context, selection_context};
+use orchestration::render_buffer::{
+    compute_buffer_layout, draw_buffer_in_split, render_buffer_in_split, resolve_cursor_fallback,
+    BufferLayoutOutput,
+};
+use orchestration::render_composite::render_composite_buffer;
 use orchestration::render_line::{render_view_lines, LastLineEnd, LineRenderInput, LineRenderOutput};
 
 use base_tokens::build_base_tokens;
@@ -87,20 +92,6 @@ const MAX_SAFE_LINE_WIDTH: usize = 10_000;
 
 
 
-/// Output of the pure layout computation phase of buffer rendering.
-/// Contains everything the drawing phase needs to produce the final frame.
-struct BufferLayoutOutput {
-    view_line_mappings: Vec<ViewLineMapping>,
-    render_output: LineRenderOutput,
-    render_area: Rect,
-    compose_layout: ComposeLayout,
-    effective_editor_bg: Color,
-    view_mode: ViewMode,
-    left_column: usize,
-    gutter_width: usize,
-    buffer_ends_with_newline: bool,
-    selection: SelectionContext,
-}
 
 
 
@@ -513,7 +504,7 @@ impl SplitRenderer {
                         }
 
                         // Render composite buffer with side-by-side panes
-                        Self::render_composite_buffer(
+                        render_composite_buffer(
                             frame,
                             layout.content_rect,
                             composite,
@@ -629,7 +620,7 @@ impl SplitRenderer {
                     .unwrap_or(&mut empty_folds);
 
                 let _render_buf_span = tracing::trace_span!("render_buffer_in_split").entered();
-                let split_view_mappings = Self::render_buffer_in_split(
+                let split_view_mappings = render_buffer_in_split(
                     frame,
                     state,
                     &split_cursors,
@@ -925,7 +916,7 @@ impl SplitRenderer {
                 .map(|vs| &mut vs.folds)
                 .unwrap_or(&mut empty_folds);
 
-            let layout_output = Self::compute_buffer_layout(
+            let layout_output = compute_buffer_layout(
                 state,
                 &split_cursors,
                 &mut viewport,
@@ -962,576 +953,6 @@ impl SplitRenderer {
     }
 
 
-    /// Render a composite buffer (side-by-side view of multiple source buffers)
-    /// Uses ViewLines for proper syntax highlighting, ANSI handling, etc.
-    #[allow(clippy::too_many_arguments)]
-    fn render_composite_buffer(
-        frame: &mut Frame,
-        area: Rect,
-        composite: &crate::model::composite_buffer::CompositeBuffer,
-        buffers: &mut HashMap<BufferId, EditorState>,
-        theme: &crate::view::theme::Theme,
-        _is_active: bool,
-        view_state: &mut crate::view::composite_view::CompositeViewState,
-        use_terminal_bg: bool,
-        show_tilde: bool,
-    ) {
-        use crate::model::composite_buffer::{CompositeLayout, RowType};
-
-        // Compute effective editor background: terminal default or theme-defined
-        let effective_editor_bg = if use_terminal_bg {
-            ratatui::style::Color::Reset
-        } else {
-            theme.editor_bg
-        };
-
-        let scroll_row = view_state.scroll_row;
-        let cursor_row = view_state.cursor_row;
-
-        // Clear the area first
-        frame.render_widget(Clear, area);
-
-        // Calculate pane widths based on layout
-        let pane_count = composite.sources.len();
-        if pane_count == 0 {
-            return;
-        }
-
-        // Extract show_separator from layout
-        let show_separator = match &composite.layout {
-            CompositeLayout::SideBySide { show_separator, .. } => *show_separator,
-            _ => false,
-        };
-
-        // Calculate pane areas
-        let separator_width = if show_separator { 1 } else { 0 };
-        let total_separators = (pane_count.saturating_sub(1)) as u16 * separator_width;
-        let available_width = area.width.saturating_sub(total_separators);
-
-        let pane_widths: Vec<u16> = match &composite.layout {
-            CompositeLayout::SideBySide { ratios, .. } => {
-                let default_ratio = 1.0 / pane_count as f32;
-                ratios
-                    .iter()
-                    .chain(std::iter::repeat(&default_ratio))
-                    .take(pane_count)
-                    .map(|r| (available_width as f32 * r).round() as u16)
-                    .collect()
-            }
-            _ => {
-                // Equal widths for stacked/unified layouts
-                let pane_width = available_width / pane_count as u16;
-                vec![pane_width; pane_count]
-            }
-        };
-
-        // Store computed pane widths in view state for cursor movement calculations
-        view_state.pane_widths = pane_widths.clone();
-
-        // Render headers first
-        let header_height = 1u16;
-        let mut x_offset = area.x;
-        for (idx, (source, &width)) in composite.sources.iter().zip(&pane_widths).enumerate() {
-            let header_area = Rect::new(x_offset, area.y, width, header_height);
-            let is_focused = idx == view_state.focused_pane;
-
-            let header_style = if is_focused {
-                Style::default()
-                    .fg(theme.tab_active_fg)
-                    .bg(theme.tab_active_bg)
-            } else {
-                Style::default()
-                    .fg(theme.tab_inactive_fg)
-                    .bg(theme.tab_inactive_bg)
-            };
-
-            let header_text = format!(" {} ", source.label);
-            let header = Paragraph::new(header_text).style(header_style);
-            frame.render_widget(header, header_area);
-
-            x_offset += width + separator_width;
-        }
-
-        // Content area (below headers)
-        let content_y = area.y + header_height;
-        let content_height = area.height.saturating_sub(header_height);
-        let visible_rows = content_height as usize;
-
-        // Render aligned rows
-        let alignment = &composite.alignment;
-        let total_rows = alignment.rows.len();
-
-        // Build ViewData and get syntax highlighting for each pane
-        // Store: (ViewLines, line->ViewLine mapping, highlight spans)
-        struct PaneRenderData {
-            lines: Vec<ViewLine>,
-            line_to_view_line: HashMap<usize, usize>,
-            highlight_spans: Vec<crate::primitives::highlighter::HighlightSpan>,
-        }
-
-        let mut pane_render_data: Vec<Option<PaneRenderData>> = Vec::new();
-
-        for (pane_idx, source) in composite.sources.iter().enumerate() {
-            if let Some(source_state) = buffers.get_mut(&source.buffer_id) {
-                // Find the first and last source lines we need for this pane
-                let visible_lines: Vec<usize> = alignment
-                    .rows
-                    .iter()
-                    .skip(scroll_row)
-                    .take(visible_rows)
-                    .filter_map(|row| row.get_pane_line(pane_idx))
-                    .map(|r| r.line)
-                    .collect();
-
-                let first_line = visible_lines.iter().copied().min();
-                let last_line = visible_lines.iter().copied().max();
-
-                if let (Some(first_line), Some(last_line)) = (first_line, last_line) {
-                    // Get byte range for highlighting
-                    let top_byte = source_state
-                        .buffer
-                        .line_start_offset(first_line)
-                        .unwrap_or(0);
-                    let end_byte = source_state
-                        .buffer
-                        .line_start_offset(last_line + 1)
-                        .unwrap_or(source_state.buffer.len());
-
-                    // Get syntax highlighting spans from the highlighter
-                    let highlight_spans = source_state.highlighter.highlight_viewport(
-                        &source_state.buffer,
-                        top_byte,
-                        end_byte,
-                        theme,
-                        1024, // highlight_context_bytes
-                    );
-
-                    // Create a temporary viewport for building view data
-                    let pane_width = pane_widths.get(pane_idx).copied().unwrap_or(80);
-                    let mut viewport =
-                        crate::view::viewport::Viewport::new(pane_width, content_height);
-                    viewport.top_byte = top_byte;
-                    viewport.line_wrap_enabled = false;
-
-                    let pane_width = pane_widths.get(pane_idx).copied().unwrap_or(80) as usize;
-                    let gutter_width = 4; // Line number width
-                    let content_width = pane_width.saturating_sub(gutter_width);
-
-                    // Build ViewData for this pane
-                    // Need enough lines to cover from first_line to last_line
-                    let lines_needed = last_line - first_line + 10;
-                    let empty_folds = FoldManager::new();
-                    let view_data = build_view_data(
-                        source_state,
-                        &viewport,
-                        None,         // No view transform
-                        80,           // estimated_line_length
-                        lines_needed, // visible_count - enough to cover the range
-                        false,        // line_wrap_enabled
-                        content_width,
-                        gutter_width,
-                        &ViewMode::Source, // Composite view uses source mode
-                        &empty_folds,
-                        theme,
-                    );
-
-                    // Build source_line -> ViewLine index mapping
-                    let mut line_to_view_line: HashMap<usize, usize> = HashMap::new();
-                    let mut current_line = first_line;
-                    for (idx, view_line) in view_data.lines.iter().enumerate() {
-                        if should_show_line_number(view_line) {
-                            line_to_view_line.insert(current_line, idx);
-                            current_line += 1;
-                        }
-                    }
-
-                    pane_render_data.push(Some(PaneRenderData {
-                        lines: view_data.lines,
-                        line_to_view_line,
-                        highlight_spans,
-                    }));
-                } else {
-                    pane_render_data.push(None);
-                }
-            } else {
-                pane_render_data.push(None);
-            }
-        }
-
-        // Now render aligned rows using ViewLines
-        for view_row in 0..visible_rows {
-            let display_row = scroll_row + view_row;
-            if display_row >= total_rows {
-                if show_tilde {
-                    // Fill with tildes for empty rows
-                    let mut x = area.x;
-                    for &width in &pane_widths {
-                        let tilde_area = Rect::new(x, content_y + view_row as u16, width, 1);
-                        let tilde =
-                            Paragraph::new("~").style(Style::default().fg(theme.line_number_fg));
-                        frame.render_widget(tilde, tilde_area);
-                        x += width + separator_width;
-                    }
-                }
-                continue;
-            }
-
-            let aligned_row = &alignment.rows[display_row];
-            let is_cursor_row = display_row == cursor_row;
-            // Get selection column range for this row (if any)
-            let selection_cols = view_state.selection_column_range(display_row);
-
-            // Determine row background based on type (selection is now character-level)
-            let row_bg = match aligned_row.row_type {
-                RowType::Addition => Some(theme.diff_add_bg),
-                RowType::Deletion => Some(theme.diff_remove_bg),
-                RowType::Modification => Some(theme.diff_modify_bg),
-                RowType::HunkHeader => Some(theme.current_line_bg),
-                RowType::Context => None,
-            };
-
-            // Compute inline diff for modified rows (to highlight changed words/characters)
-            let inline_diffs: Vec<Vec<Range<usize>>> = if aligned_row.row_type
-                == RowType::Modification
-            {
-                // Get line content from both panes
-                let mut line_contents: Vec<Option<String>> = Vec::new();
-                for (pane_idx, source) in composite.sources.iter().enumerate() {
-                    if let Some(line_ref) = aligned_row.get_pane_line(pane_idx) {
-                        if let Some(source_state) = buffers.get(&source.buffer_id) {
-                            line_contents.push(
-                                source_state
-                                    .buffer
-                                    .get_line(line_ref.line)
-                                    .map(|line| String::from_utf8_lossy(&line).to_string()),
-                            );
-                        } else {
-                            line_contents.push(None);
-                        }
-                    } else {
-                        line_contents.push(None);
-                    }
-                }
-
-                // Compute inline diff between panes (typically old vs new)
-                if line_contents.len() >= 2 {
-                    if let (Some(old_text), Some(new_text)) = (&line_contents[0], &line_contents[1])
-                    {
-                        let (old_ranges, new_ranges) = compute_inline_diff(old_text, new_text);
-                        vec![old_ranges, new_ranges]
-                    } else {
-                        vec![Vec::new(); composite.sources.len()]
-                    }
-                } else {
-                    vec![Vec::new(); composite.sources.len()]
-                }
-            } else {
-                // For non-modification rows, no inline highlighting
-                vec![Vec::new(); composite.sources.len()]
-            };
-
-            // Render each pane for this row
-            let mut x_offset = area.x;
-            for (pane_idx, (_source, &width)) in
-                composite.sources.iter().zip(&pane_widths).enumerate()
-            {
-                let pane_area = Rect::new(x_offset, content_y + view_row as u16, width, 1);
-
-                // Get horizontal scroll offset for this pane
-                let left_column = view_state
-                    .get_pane_viewport(pane_idx)
-                    .map(|v| v.left_column)
-                    .unwrap_or(0);
-
-                // Get source line for this pane
-                let source_line_opt = aligned_row.get_pane_line(pane_idx);
-
-                if let Some(source_line_ref) = source_line_opt {
-                    // Try to get ViewLine and highlight spans from pre-built data
-                    let pane_data = pane_render_data.get(pane_idx).and_then(|opt| opt.as_ref());
-                    let view_line_opt = pane_data.and_then(|data| {
-                        data.line_to_view_line
-                            .get(&source_line_ref.line)
-                            .and_then(|&idx| data.lines.get(idx))
-                    });
-                    let highlight_spans = pane_data
-                        .map(|data| data.highlight_spans.as_slice())
-                        .unwrap_or(&[]);
-
-                    let gutter_width = 4usize;
-                    let max_content_width = width.saturating_sub(gutter_width as u16) as usize;
-
-                    let is_focused_pane = pane_idx == view_state.focused_pane;
-
-                    // Determine background - cursor row highlight only on focused pane
-                    // Selection is now character-level, handled in render_view_line_content
-                    let bg = if is_cursor_row && is_focused_pane {
-                        theme.current_line_bg
-                    } else {
-                        row_bg.unwrap_or(effective_editor_bg)
-                    };
-
-                    // Selection range for this row (only for focused pane)
-                    let pane_selection_cols = if is_focused_pane {
-                        selection_cols
-                    } else {
-                        None
-                    };
-
-                    // Line number
-                    let line_num = format!("{:>3} ", source_line_ref.line + 1);
-                    let line_num_style = Style::default().fg(theme.line_number_fg).bg(bg);
-
-                    let is_cursor_pane = is_focused_pane;
-                    let cursor_column = view_state.cursor_column;
-
-                    // Get inline diff ranges for this pane
-                    let inline_ranges = inline_diffs.get(pane_idx).cloned().unwrap_or_default();
-
-                    // Determine highlight color for changed portions (brighter than line bg)
-                    let highlight_bg = match aligned_row.row_type {
-                        RowType::Deletion => Some(theme.diff_remove_highlight_bg),
-                        RowType::Addition => Some(theme.diff_add_highlight_bg),
-                        RowType::Modification => {
-                            if pane_idx == 0 {
-                                Some(theme.diff_remove_highlight_bg)
-                            } else {
-                                Some(theme.diff_add_highlight_bg)
-                            }
-                        }
-                        _ => None,
-                    };
-
-                    // Build spans using ViewLine if available (for syntax highlighting)
-                    let mut spans = vec![Span::styled(line_num, line_num_style)];
-
-                    if let Some(view_line) = view_line_opt {
-                        // Use ViewLine for syntax-highlighted content
-                        Self::render_view_line_content(
-                            &mut spans,
-                            view_line,
-                            highlight_spans,
-                            left_column,
-                            max_content_width,
-                            bg,
-                            theme,
-                            is_cursor_row && is_cursor_pane,
-                            cursor_column,
-                            &inline_ranges,
-                            highlight_bg,
-                            pane_selection_cols,
-                        );
-                    } else {
-                        // This branch should be unreachable:
-                        // - visible_lines is collected from the same range we iterate over
-                        // - If source_line_ref exists, that line was in visible_lines
-                        // - So pane_render_data exists and the line should be in the mapping
-                        // - With line_wrap disabled, each source line = one ViewLine
-                        tracing::warn!(
-                            "ViewLine missing for composite buffer: pane={}, line={}, pane_data={}",
-                            pane_idx,
-                            source_line_ref.line,
-                            pane_data.is_some()
-                        );
-                        // Graceful degradation: render empty content with background
-                        let base_style = Style::default().fg(theme.editor_fg).bg(bg);
-                        let padding = " ".repeat(max_content_width);
-                        spans.push(Span::styled(padding, base_style));
-                    }
-
-                    let line = Line::from(spans);
-                    let para = Paragraph::new(line);
-                    frame.render_widget(para, pane_area);
-                } else {
-                    // No content for this pane (padding/gap line)
-                    let is_focused_pane = pane_idx == view_state.focused_pane;
-                    // For empty lines in focused pane, show selection if entire line is selected
-                    let pane_has_selection = is_focused_pane
-                        && selection_cols
-                            .map(|(start, end)| start == 0 && end == usize::MAX)
-                            .unwrap_or(false);
-
-                    let bg = if pane_has_selection {
-                        theme.selection_bg
-                    } else if is_cursor_row && is_focused_pane {
-                        theme.current_line_bg
-                    } else {
-                        row_bg.unwrap_or(effective_editor_bg)
-                    };
-                    let style = Style::default().fg(theme.line_number_fg).bg(bg);
-
-                    // Check if cursor should be shown on this empty line
-                    let is_cursor_pane = pane_idx == view_state.focused_pane;
-                    if is_cursor_row && is_cursor_pane && view_state.cursor_column == 0 {
-                        // Show cursor on empty line
-                        let cursor_style = Style::default().fg(theme.editor_bg).bg(theme.editor_fg);
-                        let gutter_width = 4usize;
-                        let max_content_width = width.saturating_sub(gutter_width as u16) as usize;
-                        let padding = " ".repeat(max_content_width.saturating_sub(1));
-                        let line = Line::from(vec![
-                            Span::styled("    ", style),
-                            Span::styled(" ", cursor_style),
-                            Span::styled(padding, Style::default().bg(bg)),
-                        ]);
-                        let para = Paragraph::new(line);
-                        frame.render_widget(para, pane_area);
-                    } else {
-                        // Empty gap line with diff background
-                        let gap_style = Style::default().bg(bg);
-                        let empty_content = " ".repeat(width as usize);
-                        let para = Paragraph::new(empty_content).style(gap_style);
-                        frame.render_widget(para, pane_area);
-                    }
-                }
-
-                x_offset += width;
-
-                // Render separator
-                if show_separator && pane_idx < pane_count - 1 {
-                    let sep_area =
-                        Rect::new(x_offset, content_y + view_row as u16, separator_width, 1);
-                    let sep =
-                        Paragraph::new("│").style(Style::default().fg(theme.split_separator_fg));
-                    frame.render_widget(sep, sep_area);
-                    x_offset += separator_width;
-                }
-            }
-        }
-    }
-
-    /// Render ViewLine content with syntax highlighting to spans
-    #[allow(clippy::too_many_arguments)]
-    fn render_view_line_content(
-        spans: &mut Vec<Span<'static>>,
-        view_line: &ViewLine,
-        highlight_spans: &[crate::primitives::highlighter::HighlightSpan],
-        left_column: usize,
-        max_width: usize,
-        bg: Color,
-        theme: &crate::view::theme::Theme,
-        show_cursor: bool,
-        cursor_column: usize,
-        inline_ranges: &[Range<usize>],
-        highlight_bg: Option<Color>,
-        selection_cols: Option<(usize, usize)>, // (start_col, end_col) for selection
-    ) {
-        let text = &view_line.text;
-        let char_source_bytes = &view_line.char_source_bytes;
-
-        // Apply horizontal scroll and collect visible characters with styles
-        let chars: Vec<char> = text.chars().collect();
-        let mut col = 0usize;
-        let mut rendered = 0usize;
-        let mut current_span_text = String::new();
-        let mut current_style: Option<Style> = None;
-        let mut hl_cursor = 0usize;
-
-        for (char_idx, ch) in chars.iter().enumerate() {
-            let char_width = char_width(*ch);
-
-            // Skip characters before left_column
-            if col < left_column {
-                col += char_width;
-                continue;
-            }
-
-            // Stop if we've rendered enough
-            if rendered >= max_width {
-                break;
-            }
-
-            // Get source byte position for this character
-            let byte_pos = char_source_bytes.get(char_idx).and_then(|b| *b);
-
-            // Get syntax highlight color via cursor-based O(1) lookup
-            let highlight_color =
-                byte_pos.and_then(|bp| span_color_at(highlight_spans, &mut hl_cursor, bp));
-
-            // Check if this character is in an inline diff range
-            let in_inline_range = inline_ranges.iter().any(|r| r.contains(&char_idx));
-
-            // Check if this character is in selection range
-            let in_selection = selection_cols
-                .map(|(start, end)| col >= start && col < end)
-                .unwrap_or(false);
-
-            // Determine background: selection > inline diff > normal
-            let char_bg = if in_selection {
-                theme.selection_bg
-            } else if in_inline_range {
-                highlight_bg.unwrap_or(bg)
-            } else {
-                bg
-            };
-
-            // Build character style
-            let char_style = if let Some(color) = highlight_color {
-                Style::default().fg(color).bg(char_bg)
-            } else {
-                Style::default().fg(theme.editor_fg).bg(char_bg)
-            };
-
-            // Handle cursor - cursor_column is absolute position, compare directly with col
-            let final_style = if show_cursor && col == cursor_column {
-                // Invert colors for cursor
-                Style::default().fg(theme.editor_bg).bg(theme.editor_fg)
-            } else {
-                char_style
-            };
-
-            // Accumulate or flush spans based on style changes
-            if let Some(style) = current_style {
-                if style != final_style && !current_span_text.is_empty() {
-                    spans.push(Span::styled(std::mem::take(&mut current_span_text), style));
-                }
-            }
-
-            current_style = Some(final_style);
-            current_span_text.push(*ch);
-            col += char_width;
-            rendered += char_width;
-        }
-
-        // Flush remaining span
-        if !current_span_text.is_empty() {
-            if let Some(style) = current_style {
-                spans.push(Span::styled(current_span_text, style));
-            }
-        }
-
-        // Pad to fill width
-        if rendered < max_width {
-            let padding_len = max_width - rendered;
-            // cursor_column is absolute, convert to visual position for padding check
-            let cursor_visual = cursor_column.saturating_sub(left_column);
-
-            // Check if cursor is in the padding area (past end of line content)
-            if show_cursor && cursor_visual >= rendered && cursor_visual < max_width {
-                // Cursor is in padding area - render cursor as single char
-                let cursor_offset = cursor_visual - rendered;
-                let cursor_style = Style::default().fg(theme.editor_bg).bg(theme.editor_fg);
-                let normal_style = Style::default().bg(bg);
-
-                // Pre-cursor padding (if cursor is not at start of padding)
-                if cursor_offset > 0 {
-                    spans.push(Span::styled(" ".repeat(cursor_offset), normal_style));
-                }
-                // Single-char cursor
-                spans.push(Span::styled(" ", cursor_style));
-                // Post-cursor padding
-                let remaining = padding_len.saturating_sub(cursor_offset + 1);
-                if remaining > 0 {
-                    spans.push(Span::styled(" ".repeat(remaining), normal_style));
-                }
-            } else {
-                // No cursor in padding - just fill with background
-                spans.push(Span::styled(
-                    " ".repeat(padding_len),
-                    Style::default().bg(bg),
-                ));
-            }
-        }
-    }
 
 
 
@@ -1568,544 +989,6 @@ impl SplitRenderer {
 
 
 
-    fn resolve_cursor_fallback(
-        current_cursor: Option<(u16, u16)>,
-        primary_cursor_position: usize,
-        buffer_len: usize,
-        buffer_ends_with_newline: bool,
-        last_line_end: Option<LastLineEnd>,
-        lines_rendered: usize,
-        gutter_width: usize,
-    ) -> Option<(u16, u16)> {
-        if current_cursor.is_some() || primary_cursor_position != buffer_len {
-            return current_cursor;
-        }
-
-        if buffer_ends_with_newline {
-            if let Some(end) = last_line_end {
-                // When the last rendered line was the newline-terminated content
-                // line, the cursor belongs on the implicit empty line one row
-                // below.  But when the trailing empty line was already emitted
-                // by the ViewLineIterator (terminated_with_newline == false),
-                // the cursor belongs on that rendered row itself.
-                let y = if end.terminated_with_newline {
-                    end.pos.1.saturating_add(1)
-                } else {
-                    end.pos.1
-                };
-                return Some((gutter_width as u16, y));
-            }
-            return Some((gutter_width as u16, lines_rendered as u16));
-        }
-
-        last_line_end.map(|end| end.pos)
-    }
-
-    /// Pure layout computation for a buffer in a split pane.
-    /// No frame/drawing involved — produces a BufferLayoutOutput that the
-    /// drawing phase can consume.
-    #[allow(clippy::too_many_arguments)]
-    fn compute_buffer_layout(
-        state: &mut EditorState,
-        cursors: &crate::model::cursor::Cursors,
-        viewport: &mut crate::view::viewport::Viewport,
-        folds: &mut FoldManager,
-        area: Rect,
-        is_active: bool,
-        theme: &crate::view::theme::Theme,
-        lsp_waiting: bool,
-        view_mode: ViewMode,
-        compose_width: Option<u16>,
-        view_transform: Option<ViewTransformPayload>,
-        estimated_line_length: usize,
-        highlight_context_bytes: usize,
-        relative_line_numbers: bool,
-        use_terminal_bg: bool,
-        session_mode: bool,
-        software_cursor_only: bool,
-        show_line_numbers: bool,
-        highlight_current_line: bool,
-        diagnostics_inline_text: bool,
-        show_tilde: bool,
-        cell_theme_map: Option<(&mut Vec<crate::app::types::CellThemeInfo>, u16)>,
-    ) -> BufferLayoutOutput {
-        let _span = tracing::trace_span!("compute_buffer_layout").entered();
-
-        // Configure shared margin layout for this split's line number setting.
-        state.margins.configure_for_line_numbers(show_line_numbers);
-
-        // Compute effective editor background: terminal default or theme-defined
-        let effective_editor_bg = if use_terminal_bg {
-            ratatui::style::Color::Reset
-        } else {
-            theme.editor_bg
-        };
-
-        let line_wrap = viewport.line_wrap_enabled;
-
-        let overlay_count = state.overlays.all().len();
-        if overlay_count > 0 {
-            tracing::trace!("render_content: {} overlays present", overlay_count);
-        }
-
-        let visible_count = viewport.visible_line_count();
-
-        let buffer_len = state.buffer.len();
-        let byte_offset_mode = state.buffer.line_count().is_none();
-        let estimated_lines = if byte_offset_mode {
-            // In byte offset mode, gutter shows byte offsets, so size the gutter
-            // for the largest byte offset (file size)
-            buffer_len.max(1)
-        } else {
-            state.buffer.line_count().unwrap_or(1)
-        };
-        state
-            .margins
-            .update_width_for_buffer(estimated_lines, show_line_numbers);
-        let gutter_width = state.margins.left_total_width();
-
-        let compose_layout = calculate_compose_layout(area, &view_mode, compose_width);
-        let render_area = compose_layout.render_area;
-
-        // Clone view_transform so we can reuse it if scrolling triggers a rebuild
-        let view_transform_for_rebuild = view_transform.clone();
-
-        let view_data = {
-            let _span = tracing::trace_span!("build_view_data").entered();
-            build_view_data(
-                state,
-                viewport,
-                view_transform,
-                estimated_line_length,
-                visible_count,
-                line_wrap,
-                render_area.width as usize,
-                gutter_width,
-                &view_mode,
-                folds,
-                theme,
-            )
-        };
-
-        // Same-buffer scroll sync: if the sync code flagged this viewport to
-        // scroll to the end, apply it now using the view lines we just built.
-        let sync_scrolled = if viewport.sync_scroll_to_end {
-            viewport.sync_scroll_to_end = false;
-            viewport.scroll_to_end_of_view(&view_data.lines)
-        } else {
-            false
-        };
-
-        // If the sync adjustment changed top_byte, rebuild view_data before
-        // ensure_visible_in_layout runs (so it sees the correct view lines).
-        let (view_data, view_transform_for_rebuild) = if sync_scrolled {
-            viewport.top_view_line_offset = 0;
-            let rebuilt = build_view_data(
-                state,
-                viewport,
-                view_transform_for_rebuild,
-                estimated_line_length,
-                visible_count,
-                line_wrap,
-                render_area.width as usize,
-                gutter_width,
-                &view_mode,
-                folds,
-                theme,
-            );
-            viewport.scroll_to_end_of_view(&rebuilt.lines);
-            (rebuilt, None)
-        } else {
-            (view_data, Some(view_transform_for_rebuild))
-        };
-
-        // Ensure cursor is visible using Layout-aware check (handles virtual lines)
-        let primary = *cursors.primary();
-        let scrolled = viewport.ensure_visible_in_layout(&view_data.lines, &primary, gutter_width);
-
-        // If we scrolled, rebuild view_data from the new top_byte and then re-run the
-        // layout-aware check so that top_view_line_offset is correct for the rebuilt data.
-        let view_data = if scrolled {
-            if let Some(vt) = view_transform_for_rebuild {
-                viewport.top_view_line_offset = 0;
-                let rebuilt = build_view_data(
-                    state,
-                    viewport,
-                    vt,
-                    estimated_line_length,
-                    visible_count,
-                    line_wrap,
-                    render_area.width as usize,
-                    gutter_width,
-                    &view_mode,
-                    folds,
-                    theme,
-                );
-                let _ = viewport.ensure_visible_in_layout(&rebuilt.lines, &primary, gutter_width);
-                rebuilt
-            } else {
-                view_data
-            }
-        } else {
-            view_data
-        };
-
-        let view_anchor = calculate_view_anchor(&view_data.lines, viewport.top_byte);
-
-        let selection = selection_context(state, cursors);
-
-        tracing::trace!(
-            "Rendering buffer with {} cursors at positions: {:?}, primary at {}, is_active: {}, buffer_len: {}",
-            selection.cursor_positions.len(),
-            selection.cursor_positions,
-            selection.primary_cursor_position,
-            is_active,
-            state.buffer.len()
-        );
-
-        if !selection.cursor_positions.is_empty()
-            && !selection
-                .cursor_positions
-                .contains(&selection.primary_cursor_position)
-        {
-            tracing::warn!(
-                "Primary cursor position {} not found in cursor_positions list: {:?}",
-                selection.primary_cursor_position,
-                selection.cursor_positions
-            );
-        }
-
-        let adjusted_visible_count = fold_adjusted_visible_count(
-            &state.buffer,
-            &state.marker_list,
-            folds,
-            viewport.top_byte,
-            visible_count,
-        );
-
-        // Populate line cache to ensure chunks are loaded for rendering.
-        // For small files this also builds the line index; for large files
-        // it just loads the needed chunks from disk.
-        let _ = state
-            .buffer
-            .populate_line_cache(viewport.top_byte, adjusted_visible_count);
-
-        let viewport_start = viewport.top_byte;
-        let viewport_end = calculate_viewport_end(
-            state,
-            viewport_start,
-            estimated_line_length,
-            adjusted_visible_count,
-        );
-
-        let decorations = decoration_context(
-            state,
-            viewport_start,
-            viewport_end,
-            selection.primary_cursor_position,
-            folds,
-            theme,
-            highlight_context_bytes,
-            &view_mode,
-            diagnostics_inline_text,
-            &view_data.lines,
-        );
-
-        let calculated_offset = viewport.top_view_line_offset;
-
-        tracing::trace!(
-            top_byte = viewport.top_byte,
-            top_view_line_offset = viewport.top_view_line_offset,
-            calculated_offset,
-            view_data_lines = view_data.lines.len(),
-            "view line offset calculation"
-        );
-        let (view_lines_to_render, adjusted_view_anchor) =
-            if calculated_offset > 0 && calculated_offset < view_data.lines.len() {
-                let sliced = &view_data.lines[calculated_offset..];
-                let adjusted_anchor = calculate_view_anchor(sliced, viewport.top_byte);
-                (sliced, adjusted_anchor)
-            } else {
-                (&view_data.lines[..], view_anchor)
-            };
-
-        // Use provided cell theme map or a temporary dummy
-        let mut dummy_map = Vec::new();
-        let (map_ref, sw) = match cell_theme_map {
-            Some((map, w)) => (map, w),
-            None => (&mut dummy_map, 0u16),
-        };
-
-        let render_output = render_view_lines(LineRenderInput {
-            state,
-            theme,
-            view_lines: view_lines_to_render,
-            view_anchor: adjusted_view_anchor,
-            render_area,
-            gutter_width,
-            selection: &selection,
-            decorations: &decorations,
-            visible_line_count: visible_count,
-            lsp_waiting,
-            is_active,
-            line_wrap,
-            estimated_lines,
-            left_column: viewport.left_column,
-            relative_line_numbers,
-            session_mode,
-            software_cursor_only,
-            show_line_numbers,
-            byte_offset_mode,
-            show_tilde,
-            highlight_current_line,
-            cell_theme_map: map_ref,
-            screen_width: sw,
-        });
-
-        let view_line_mappings = render_output.view_line_mappings.clone();
-
-        let buffer_ends_with_newline = if !state.buffer.is_empty() {
-            let last_char = state.get_text_range(state.buffer.len() - 1, state.buffer.len());
-            last_char == "\n"
-        } else {
-            false
-        };
-
-        BufferLayoutOutput {
-            view_line_mappings,
-            render_output,
-            render_area,
-            compose_layout,
-            effective_editor_bg,
-            view_mode,
-            left_column: viewport.left_column,
-            gutter_width,
-            buffer_ends_with_newline,
-            selection,
-        }
-    }
-
-    /// Draw a buffer into a frame using pre-computed layout output.
-    #[allow(clippy::too_many_arguments)]
-    fn draw_buffer_in_split(
-        frame: &mut Frame,
-        state: &EditorState,
-        cursors: &crate::model::cursor::Cursors,
-        layout_output: BufferLayoutOutput,
-        event_log: Option<&mut EventLog>,
-        area: Rect,
-        is_active: bool,
-        theme: &crate::view::theme::Theme,
-        ansi_background: Option<&AnsiBackground>,
-        background_fade: f32,
-        hide_cursor: bool,
-        software_cursor_only: bool,
-        rulers: &[usize],
-        compose_column_guides: Option<Vec<u16>>,
-    ) {
-        let render_area = layout_output.render_area;
-        let effective_editor_bg = layout_output.effective_editor_bg;
-        let gutter_width = layout_output.gutter_width;
-        let starting_line_num = 0; // used only for background offset
-
-        render_compose_margins(
-            frame,
-            area,
-            &layout_output.compose_layout,
-            &layout_output.view_mode,
-            theme,
-            effective_editor_bg,
-        );
-
-        let mut lines = layout_output.render_output.lines;
-        let background_x_offset = layout_output.left_column;
-
-        if let Some(bg) = ansi_background {
-            apply_background_to_lines(
-                &mut lines,
-                render_area.width,
-                bg,
-                effective_editor_bg,
-                theme.editor_fg,
-                background_fade,
-                background_x_offset,
-                starting_line_num,
-            );
-        }
-
-        frame.render_widget(Clear, render_area);
-        let editor_block = Block::default()
-            .borders(Borders::NONE)
-            .style(Style::default().bg(effective_editor_bg));
-        frame.render_widget(Paragraph::new(lines).block(editor_block), render_area);
-
-        let cursor = Self::resolve_cursor_fallback(
-            layout_output.render_output.cursor,
-            layout_output.selection.primary_cursor_position,
-            state.buffer.len(),
-            layout_output.buffer_ends_with_newline,
-            layout_output.render_output.last_line_end,
-            layout_output.render_output.content_lines_rendered,
-            gutter_width,
-        );
-
-        let cursor_screen_pos = if is_active && state.show_cursors && !hide_cursor {
-            cursor.map(|(cx, cy)| {
-                let screen_x = render_area.x.saturating_add(cx);
-                let max_y = render_area.height.saturating_sub(1);
-                let screen_y = render_area.y.saturating_add(cy.min(max_y));
-                (screen_x, screen_y)
-            })
-        } else {
-            None
-        };
-
-        // Render config-based vertical rulers
-        if !rulers.is_empty() {
-            let ruler_cols: Vec<u16> = rulers.iter().map(|&r| r as u16).collect();
-            render_ruler_bg(
-                frame,
-                &ruler_cols,
-                theme.ruler_bg,
-                render_area,
-                gutter_width,
-                layout_output.render_output.content_lines_rendered,
-                layout_output.left_column,
-            );
-        }
-
-        // Render compose column guides
-        if let Some(guides) = compose_column_guides {
-            let guide_style = Style::default()
-                .fg(theme.line_number_fg)
-                .add_modifier(Modifier::DIM);
-            render_column_guides(
-                frame,
-                &guides,
-                guide_style,
-                render_area,
-                gutter_width,
-                layout_output.render_output.content_lines_rendered,
-                0,
-            );
-        }
-
-        if let Some((screen_x, screen_y)) = cursor_screen_pos {
-            frame.set_cursor_position((screen_x, screen_y));
-
-            // When software_cursor_only the backend has no hardware cursor, so
-            // ensure the cell at the cursor position always has REVERSED style.
-            // This covers all edge cases (end-of-line, empty buffer, newline
-            // positions) where the per-character REVERSED styling from
-            // compute_char_style may not have been applied.
-            if software_cursor_only {
-                let buf = frame.buffer_mut();
-                let area = buf.area;
-                if screen_x < area.x + area.width && screen_y < area.y + area.height {
-                    let cell = &mut buf[(screen_x, screen_y)];
-                    // Only override empty / default-background cells to avoid
-                    // double-reversing cells that already got software cursor
-                    // styling in render_view_lines.
-                    if !cell.modifier.contains(Modifier::REVERSED) {
-                        cell.set_char(' ');
-                        cell.fg = theme.editor_fg;
-                        cell.bg = theme.editor_bg;
-                        cell.modifier.insert(Modifier::REVERSED);
-                    }
-                }
-            }
-
-            if let Some(event_log) = event_log {
-                let cursor_pos = cursors.primary().position;
-                let buffer_len = state.buffer.len();
-                event_log.log_render_state(cursor_pos, screen_x, screen_y, buffer_len);
-            }
-        }
-    }
-
-    /// Render a single buffer in a split pane (convenience wrapper).
-    /// Calls compute_buffer_layout then draw_buffer_in_split.
-    /// Returns the view line mappings for mouse click handling.
-    #[allow(clippy::too_many_arguments)]
-    fn render_buffer_in_split(
-        frame: &mut Frame,
-        state: &mut EditorState,
-        cursors: &crate::model::cursor::Cursors,
-        viewport: &mut crate::view::viewport::Viewport,
-        folds: &mut FoldManager,
-        event_log: Option<&mut EventLog>,
-        area: Rect,
-        is_active: bool,
-        theme: &crate::view::theme::Theme,
-        ansi_background: Option<&AnsiBackground>,
-        background_fade: f32,
-        lsp_waiting: bool,
-        view_mode: ViewMode,
-        compose_width: Option<u16>,
-        compose_column_guides: Option<Vec<u16>>,
-        view_transform: Option<ViewTransformPayload>,
-        estimated_line_length: usize,
-        highlight_context_bytes: usize,
-        _buffer_id: BufferId,
-        hide_cursor: bool,
-        relative_line_numbers: bool,
-        use_terminal_bg: bool,
-        session_mode: bool,
-        software_cursor_only: bool,
-        rulers: &[usize],
-        show_line_numbers: bool,
-        highlight_current_line: bool,
-        diagnostics_inline_text: bool,
-        show_tilde: bool,
-        cell_theme_map: &mut Vec<crate::app::types::CellThemeInfo>,
-        screen_width: u16,
-    ) -> Vec<ViewLineMapping> {
-        let layout_output = Self::compute_buffer_layout(
-            state,
-            cursors,
-            viewport,
-            folds,
-            area,
-            is_active,
-            theme,
-            lsp_waiting,
-            view_mode.clone(),
-            compose_width,
-            view_transform,
-            estimated_line_length,
-            highlight_context_bytes,
-            relative_line_numbers,
-            use_terminal_bg,
-            session_mode,
-            software_cursor_only,
-            show_line_numbers,
-            highlight_current_line,
-            diagnostics_inline_text,
-            show_tilde,
-            Some((cell_theme_map, screen_width)),
-        );
-
-        let view_line_mappings = layout_output.view_line_mappings.clone();
-
-        Self::draw_buffer_in_split(
-            frame,
-            state,
-            cursors,
-            layout_output,
-            event_log,
-            area,
-            is_active,
-            theme,
-            ansi_background,
-            background_fade,
-            hide_cursor,
-            software_cursor_only,
-            rulers,
-            compose_column_guides,
-        );
-
-        view_line_mappings
-    }
 
 
 }
@@ -2353,7 +1236,7 @@ mod tests {
     #[test]
     fn cursor_after_newline_places_on_next_line() {
         let (output, buffer_len, buffer_newline, cursor_pos) = render_output_for("abc\n", 4);
-        let cursor = SplitRenderer::resolve_cursor_fallback(
+        let cursor = resolve_cursor_fallback(
             output.cursor,
             cursor_pos,
             buffer_len,
@@ -2368,7 +1251,7 @@ mod tests {
     #[test]
     fn cursor_at_end_without_newline_stays_on_line() {
         let (output, buffer_len, buffer_newline, cursor_pos) = render_output_for("abc", 3);
-        let cursor = SplitRenderer::resolve_cursor_fallback(
+        let cursor = resolve_cursor_fallback(
             output.cursor,
             cursor_pos,
             buffer_len,
@@ -2463,7 +1346,7 @@ mod tests {
             all_cursors
         );
 
-        let final_cursor = SplitRenderer::resolve_cursor_fallback(
+        let final_cursor = resolve_cursor_fallback(
             output.cursor,
             cursor_pos,
             buffer_len,
@@ -2690,7 +1573,7 @@ mod tests {
             output.cursor
         );
 
-        let final_cursor = SplitRenderer::resolve_cursor_fallback(
+        let final_cursor = resolve_cursor_fallback(
             output.cursor,
             cursor_pos,
             buffer_len,
